+++ src/order/application/use-cases/create-order.use-case.spec.ts
@@ -2,6 +2,7 @@ import { Test } from "@nestjs/testing";
 import { CreateOrderUseCase } from "./create-order.use-case";
 import { OrderStatus } from "@/order/domain/entities/order.entitiy";
 import { v4 as uuidv4 } from "uuid";
+import { CreateOrderDomainService } from "@/order/domain/services/create-order.service";
 
 describe("CreateOrderUseCase", () => {
   let useCase: CreateOrderUseCase;
@@ -20,6 +21,7 @@ describe("CreateOrderUseCase", () => {
     const module = await Test.createTestingModule({
       providers: [
         CreateOrderUseCase,
+        CreateOrderDomainService,
         {
           provide: "OrderRepositoryInterface",
           useValue: orderRepository,
