+++ src/order/application/use-cases/create-order.use-case.ts
@@ -1,8 +1,8 @@
 import { Inject, Injectable } from "@nestjs/common";
-import { Order, OrderStatus } from "@/order/domain/entities/order.entitiy";
+import { Order } from "@/order/domain/entities/order.entitiy";
 import { OrderRepositoryInterface } from "@/order/domain/interfaces/order.repository.interface";
-import { OrderItem } from "@/order/domain/entities/order-item.entity";
 import { OrderItemRepositoryInterface } from "@/order/domain/interfaces/order-item.repository.interface";
+import { CreateOrderDomainService } from "@/order/domain/services/create-order.service";
 
 export interface CreateOrderUseCaseCommand {
   userId: string;
@@ -24,7 +24,8 @@ export class CreateOrderUseCase {
     @Inject("OrderRepositoryInterface")
     private readonly orderRepository: OrderRepositoryInterface,
     @Inject("OrderItemRepositoryInterface")
-    private readonly orderItemRepository: OrderItemRepositoryInterface
+    private readonly orderItemRepository: OrderItemRepositoryInterface,
+    private readonly createOrderDomainService: CreateOrderDomainService
   ) {}
 
   async execute(
@@ -32,30 +33,18 @@ export class CreateOrderUseCase {
   ): Promise<CreateOrderUseCaseResult> {
     const { userId, idempotencyKey, items } = command;
 
-    const order = Order.create({
+    const order = await this.createOrderDomainService.createOrder({
       userId,
-      totalPrice: 0,
-      discountPrice: 0,
-      finalPrice: 0,
-      status: OrderStatus.PENDING,
       idempotencyKey,
+      items,
     });
 
-    const orderItems = items.map((item) =>
-      OrderItem.create({
-        ...item,
-        orderId: order.id,
-      })
-    );
-
-    order.initOrderItems(orderItems);
-
-    await Promise.all([
-      this.orderRepository.save(order),
-      ...orderItems.map((orderItem) =>
+    await this.orderRepository.save(order);
+    await Promise.all(
+      order.orderItems.map((orderItem) =>
         this.orderItemRepository.save(orderItem)
-      ),
-    ]);
+      )
+    );
 
     return {
       order,
